requirements.txt
# TODO: Add project dependencies

fastapi
uvicorn[standard]

python-multipart

openai

python-dotenv

supabase

python-telegram-bot==20.3
